Laravel Admin Module Complete for microservice

+++ app/Http/Controllers/RoleController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\RoleResource;
 use App\Models\Role;
+use DB;
+use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -13,7 +16,8 @@ class RoleController extends Controller
      */
     public function index()
     {
-        return Role::all();
+        Gate::authorize('view', 'roles');
+        return RoleResource::collection(Role::all());
     }
 
     /**
@@ -21,8 +25,19 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
         $role = Role::create($request->only('name'));
-        return response(['role' => $role, 'message' => 'role created successfully!!'],Response::HTTP_CREATED);
+
+        if($permissions = $request->input('permissions')){
+            foreach($permissions as $permission_id){
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     /**
@@ -30,7 +45,8 @@ class RoleController extends Controller
      */
     public function show(string $id)
     {
-        return Role::find($id);
+        Gate::authorize('view', 'roles');
+        return new RoleResource(Role::find($id));
     }
 
     /**
@@ -38,11 +54,23 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('edit', 'roles');
         $role = Role::find($id);
 
         $role->update($request->only('name'));
 
-        return response(['role' => $role, 'message' => 'role updated successfully!!'], Response::HTTP_ACCEPTED);
+        DB::table('role_permission')->where('role_id', $role->id)->delete();
+
+        if($permissions = $request->input('permissions')){
+            foreach($permissions as $permission_id){
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
     /**
@@ -50,6 +78,8 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('edit', 'roles');
+        DB::table('role_permission')->where('role_id', $id)->delete();
         Role::destroy($id);
 
         return response(['message' => 'role deleted successfully!'], Response::HTTP_NO_CONTENT);
